telegram bot file separeting
